# swagger.yaml
basePath: /api
definitions:
...
    GoogleAuthenticateRequest:
        properties:
            code:
                type: string
                x-go-name: Code
        required:
            - code
        type: object
        x-go-package: github.com/beecool-cocktail/application-backend/viewmodels
...
    LogoutRequest:
        properties:
            user_id:
                format: int64
                type: integer
                x-go-name: UserID
        required:
            - user_id
        type: object
        x-go-package: github.com/beecool-cocktail/application-backend/viewmodels
...
    UpdateUserAvatarRequest:
        properties:
            coordinate:
                description: 座標 [左上XY, 右下XY]
                items:
                    $ref: '#/definitions/Coordinate'
                type: array
                x-go-name: Coordinate
            crop_avatar:
                description: 裁切後照片
                type: string
                x-go-name: CropAvatar
            origin_avatar:
                description: 原始照片
                type: string
                x-go-name: OriginAvatar
            rotation:
                description: 旋轉角度
                format: float
                type: number
                x-go-name: Rotation
        required:
            - crop_avatar
            - coordinate
            - rotation
        type: object
        x-go-package: github.com/beecool-cocktail/application-backend/viewmodels
...
host: localhost:6969
info:
    description: This is api server
    title: api server
    version: "0.1"
paths:
    /auth/google-authenticate:
        post:
            description: Use Code to exchange access token.
            operationId: googleAuthenticateRequest
            parameters:
                - in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/GoogleAuthenticateRequest'
            responses:
                "201":
                    $ref: '#/responses/googleAuthenticateResponse'
            summary: Get access token.
            tags:
                - login
...
    /auth/logout:
        post:
            description: make token invalid.
            operationId: logoutRequest
            parameters:
                - in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/LogoutRequest'
            responses:
                "200":
                    description: success
            summary: User logout.
            tags:
                - login
...
    /users/{id}:
        get:
            description: Get other user id, name, email, numberOfPost, numberOfCollection and photo.
            operationId: getOtherUserInfo
            parameters:
                - example: 1
                  in: path
                  name: id
                  required: true
                  type: integer
            responses:
                "200":
                    $ref: '#/responses/getOtherUserInfoResponse'
            security:
                - Bearer:
                    - apiKey
            summary: Get other user information.
            tags:
                - user
    /users/{id}/cocktails:
        get:
            description: Get other user cocktail list order by create date.
            operationId: otherCocktailList
            parameters:
                - example: 1
                  in: path
                  name: id
                  required: true
                  type: integer
            responses:
                "200":
                    $ref: '#/responses/getOtherCocktailListResponse'
            security:
                - Bearer:
                    - apiKey
            summary: Get other user cocktail list
            tags:
                - user
    /users/{id}/favorite-cocktails:
        get:
            description: Get other user favorite cocktail article list.
            operationId: getOtherUserFavoriteList
            parameters:
                - example: 123456
                  in: path
                  name: id
                  required: true
                  type: integer
            responses:
                "200":
                    $ref: '#/responses/getUserFavoriteCocktailListResponse'
            security:
                - Bearer:
                    - apiKey
            summary: Get other user favorite cocktail article list.
            tags:
                - user
    /users/current:
        get:
            description: Get user id, name, email, numberOfPost, numberOfCollection and photo.
            operationId: info
            responses:
                "200":
                    $ref: '#/responses/getUserInfoResponse'
            security:
                - Bearer:
                    - apiKey
            summary: Get user information.
            tags:
                - user
    /users/current/avatar:
        delete:
            description: Delete user avatar.
            operationId: deleteUserAvatar
            responses:
                "200":
                    description: success
            security:
                - Bearer:
                    - apiKey
            summary: Delete user avatar.
            tags:
                - user
        put:
            description: Edit user avatar.
            operationId: updateUserAvatarRequest
            parameters:
                - in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/UpdateUserAvatarRequest'
            responses:
                "200":
                    description: success
            security:
                - Bearer:
                    - apiKey
            summary: Edit user avatar.
            tags:
                - user
    /users/current/cocktails:
        get:
            description: Get self cocktail list order by create date.
            operationId: selfCocktailList
            responses:
                "200":
                    $ref: '#/responses/getSelfCocktailListResponse'
            security:
                - Bearer:
                    - apiKey
            summary: Get self cocktail list
            tags:
                - user
    /users/current/favorite-cocktails:
        get:
            description: Get current user favorite cocktail article list.
            operationId: getUserFavoriteList
            responses:
                "200":
                    $ref: '#/responses/getUserFavoriteCocktailListResponse'
            security:
                - Bearer:
                    - apiKey
            summary: Get current user favorite cocktail article list.
            tags:
                - user
        post:
            description: Add cocktail article to favorite list.
            operationId: collectArticleRequest
            parameters:
                - in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/CollectArticleRequest'
            responses:
                "201":
                    description: success
            security:
                - Bearer:
                    - apiKey
            summary: Add cocktail article to favorite list.
            tags:
                - user
    /users/current/favorite-cocktails/{id}:
        delete:
            description: Remove cocktail article from favorite list.
            operationId: removeCollectionArticle
            parameters:
                - example: 123456
                  in: path
                  name: id
                  required: true
                  type: integer
            responses:
                "200":
                    $ref: '#/responses/deleteFavoriteCocktailResponse'
            security:
                - Bearer:
                    - apiKey
            summary: Remove cocktail article from favorite list.
            tags:
                - user
    /users/current/info:
        put:
            description: Edit user name and collection of publicity status.
            operationId: updateUserInfoRequest
            parameters:
                - in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/UpdateUserInfoRequest'
            responses:
                "200":
                    $ref: '#/responses/updateUserPhotoResponse'
            security:
                - Bearer:
                    - apiKey
            summary: Edit user information.
            tags:
                - user
responses:
...
    getOtherCocktailListResponse:
        description: ""
        schema:
            properties:
                data:
                    $ref: '#/definitions/GetOtherCocktailListResponse'
                error_code:
                    type: string
                    x-go-name: ErrorCode
                error_message:
                    type: string
                    x-go-name: ErrorMessage
            required:
                - error_code
                - error_message
                - data
            type: object
    getOtherUserInfoResponse:
        description: ""
        schema:
            properties:
                data:
                    $ref: '#/definitions/GetOtherUserInfoResponse'
                error_code:
                    type: string
                    x-go-name: ErrorCode
                error_message:
                    type: string
                    x-go-name: ErrorMessage
            required:
                - error_code
                - error_message
                - data
            type: object
    getSelfCocktailListResponse:
        description: ""
        schema:
            properties:
                data:
                    $ref: '#/definitions/GetSelfCocktailListResponse'
                error_code:
                    type: string
                    x-go-name: ErrorCode
                error_message:
                    type: string
                    x-go-name: ErrorMessage
            required:
                - error_code
                - error_message
                - data
            type: object
    getUserFavoriteCocktailListResponse:
        description: ""
        schema:
            properties:
                data:
                    $ref: '#/definitions/GetUserFavoriteCocktailListResponse'
                error_code:
                    type: string
                    x-go-name: ErrorCode
                error_message:
                    type: string
                    x-go-name: ErrorMessage
            required:
                - error_code
                - error_message
                - data
            type: object
    getUserInfoResponse:
        description: ""
        schema:
            properties:
                data:
                    $ref: '#/definitions/GetUserInfoResponse'
                error_code:
                    type: string
                    x-go-name: ErrorCode
                error_message:
                    type: string
                    x-go-name: ErrorMessage
            required:
                - error_code
                - error_message
                - data
            type: object
    googleAuthenticateResponse:
        description: ""
        schema:
            properties:
                data:
                    $ref: '#/definitions/GoogleAuthenticateResponse'
                error_code:
                    type: string
                    x-go-name: ErrorCode
                error_message:
                    type: string
                    x-go-name: ErrorMessage
            required:
                - error_code
                - error_message
                - data
            type: object
...
schemes:
    - http
    - https
securityDefinitions:
    Bearer:
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"